Generate Artifacts Outside the Tags Context (#15)
This pull request introduces changes to improve the handling of GitHub
Actions workflows by adding support for tag-based operations and
refactoring the configuration setup. Additionally, it includes updates
to testing configurations and test cases to align with these changes.

### GitHub Actions Workflow Enhancements:
*
[`src/configuration.ts`](diffhunk://#diff-0931cf958317ce57b36bdf11695eaa9ce4ebeec0f9ae067343159a85d75732dfL7-R10):
Added a new `env` parameter to the `Configuration` class and a method
`isTag` to determine if the workflow is triggered by a tag reference.
[[1]](diffhunk://#diff-0931cf958317ce57b36bdf11695eaa9ce4ebeec0f9ae067343159a85d75732dfL7-R10)
[[2]](diffhunk://#diff-0931cf958317ce57b36bdf11695eaa9ce4ebeec0f9ae067343159a85d75732dfR19-R22)
*
[`src/main.ts`](diffhunk://#diff-4fab5baaca5c14d2de62d8d2fceef376ddddcc8e9509d86cfa5643f51b89ce3dL9-R20):
Updated the `main` function to conditionally execute temporary branch
creation and deletion based on the `isTag` property.
*
[`src/model/artifacts.ts`](diffhunk://#diff-5595f77bb2b2f49522043f8dcd486390018618b047f219418459269c756ecc03L19-R21):
Modified the `Artifacts` class to conditionally execute tag-related
tasks (`collect` and `move`) based on the `isTag` property.

### Testing Updates:
*
[`tests/unit/model/artifacts.test.ts`](diffhunk://#diff-7038a68cfc252210ef35773ee922184899920f50ea4663f60096442511db507eR138-R180):
Added a new test case to verify that tag-related tasks are skipped when
the action is not triggered by a tag. Updated the `configuration`
function to accept an optional `env` parameter for testing purposes.

### Configuration and File Structure Adjustments:
*
[`jest.config.ts`](diffhunk://#diff-860bd1f15d1e0bafcfc6f62560524f588e6d6bf56d4ab1b0f6f8146461558730L7-R10):
Updated paths in `moduleNameMapper` and `setupFilesAfterEnv` to reflect
the new file structure after renaming `tests/jest.config.ts` to
`jest.config.ts`.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L33-R33):
Simplified the Jest configuration in the `test` script to use the
default `jest.config.ts`.
*
[`tests/tsconfig.json`](diffhunk://#diff-9ef891bf14629264f353ee51ca7dad4267563a5d90ad7e3c62030eba140bf5c9L12-R12):
Adjusted the `include` paths to match the updated file structure.